+++ Source/DocButton.cpp
@@ -24,7 +24,7 @@ DocButton::DocButton()
 	GetGraphSizeF(hImageNormal_, &rect_.width, &rect_.height);
 
 	rect_.x = Screen::WIDTH - ((rect_.width + BUTTON_PADDING) * 1);
-	rect_.y = BUTTON_PADDING;
+	rect_.y = static_cast<float>(BUTTON_PADDING);
 
 	SetHintText("ÉhÉLÉÖÉÅÉìÉgÇå©ÇÈ");
 }
@@ -43,12 +43,12 @@ void DocButton::OnDraw(const bool _onTouching, const bool _isPushing)
 	{
 		SetDrawBlendMode(DX_BLENDMODE_ALPHA, 100);
 	}
-	DrawGraph(rect_.x, rect_.y, hImage_, TRUE);
+	DrawGraph(static_cast<int>(rect_.x), static_cast<int>(rect_.y), hImage_, TRUE);
 
 	if (_isPushing)
 	{
 		SetDrawBlendMode(DX_BLENDMODE_ADD, 100);
-		DrawGraph(rect_.x, rect_.y, hImage_, TRUE);
+		DrawGraph(static_cast<int>(rect_.x), static_cast<int>(rect_.y), hImage_, TRUE);
 	}
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 }
