+++ Source/FinalTestButton.cpp
@@ -34,19 +34,19 @@ void FinalTestButton::OnDraw(const bool _onTouching, const bool _isPushing)
 {
 	if (_onTouching)
 	{
-		DrawGraph(rect_.x, rect_.y, hImage_, TRUE);
+		DrawGraph(static_cast<int>(rect_.x), static_cast<int>(rect_.y), hImage_, TRUE);
 		SetDrawBlendMode(DX_BLENDMODE_ADD, 100);
 	}
 	else
 	{
 		SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 	}
-	DrawGraph(rect_.x, rect_.y, hImage_, TRUE);
+	DrawGraph(static_cast<int>(rect_.x), static_cast<int>(rect_.y), hImage_, TRUE);
 
 	if (_isPushing)
 	{
 		SetDrawBlendMode(DX_BLENDMODE_ADD, 100);
-		DrawGraph(rect_.x, rect_.y, hImage_, TRUE);
+		DrawGraph(static_cast<int>(rect_.x), static_cast<int>(rect_.y), hImage_, TRUE);
 	}
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 }
